requirements.txt
openai-whisper
torch
ffmpeg-python 
setuptools
numpy<2.0
